catch event handling free() bug
The event handling component assumes wrongly that
every scheduled event is a "one shot event" and freed
the wrapper structure. But this was/is not true at all.
The timer handler are one shot events and freed and later
re-added to the event mechanism.

On the other side the raw socket event fd is a daemon lifetime
fd - with additional lifetime wrapper. It is wrong to free this
structure.

The patch now handles this by add an additional flag to make the
behaviour known:

o one-shot
o lifetime

See event.h for the exact definition.