onvert (incoming) packet handling complete to use o_bufs
Until now the packet handling system was based on a
simple data pointer. The current employed concept
adjust that and uses the newly created packet abstraction
struct o_buf - which should[TM] be similar to the m_buf
and sk_buf wrapper.

The current wrapper functionality is really limited - but
heh: start with the simple solution and drill out if required.