determine interface for incoming packet
At the stage of incoming HELLO packets the current
interface is an index into the interface area. The current
code uses the previously gathered interface id to get the interface
of the configured routing domain.

In the case that this interface was not configured properly for
routing an warning message is displayed.