routine to start timeout handling for neighbors
Add the required functionality to detect new neighbors
and to process already known neighbors. A lot of changes
concerns the timer handling functionality. At least a more
abstract concept could be introduced to nicely simplify the
process of the different timeout handling routines.

At the current stage the daemon send and receive new HELLO messages
and process then in a very redumentary sense.